+++ lib/mongoid/config.rb
@@ -216,5 +216,17 @@ module Mongoid
     def time_zone
       use_utc? ? "UTC" : ::Time.zone
     end
+
+    # Is the application running under passenger?
+    #
+    # @example Is the application using passenger?
+    #   config.running_with_passenger?
+    #
+    # @return [ true, false ] If the app is deployed on Passenger.
+    #
+    # @since 3.0.11
+    def running_with_passenger?
+      @running_with_passenger ||= defined?(PhusionPassenger)
+    end
   end
 end